@@ -2,26 +2,26 @@
   "name": "frontend",
   "version": "0.1.0",
   "private": true,
-  "proxy": "http://localhost:5000",
+  "proxy": "http://localhost:3333",
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "dateformat": "^3.0.3",
-    "lodash": "^4.17.15",
-    "node-sass": "^4.13.1",
-    "react": "^16.12.0",
-    "react-date-picker": "^7.10.0",
-    "react-day-picker": "^7.4.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.4.1",
-    "redux": "^4.0.5",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "dateformat": "^5.0.3",
+    "lodash": "4.17.21",
+    "react": "^19.1.0",
+    "react-date-picker": "^11.0.0",
+    "react-day-picker": "^9.8.0",
+    "react-dom": "^19.1.0",
+    "react-redux": "^9.2.0",
+    "react-router-dom": "^7.6.3",
+    "react-scripts": "^5.0.1",
+    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
     "redux-persist": "^6.0.0",
-    "redux-persist-transform-filter": "^0.0.20"
+    "redux-persist-transform-filter": "^0.0.22",
+    "redux-thunk": "^3.1.0",
+    "sass": "^1.89.2"
   },
   "scripts": {
     "start": "react-scripts start",
