@@ -1,9 +1,10 @@
-module.exports = {
+export default {
     mongoURI: process.env.MONGO_URI,
     secretOrKey: process.env.SECRET_OR_KEY,
     awsBucketAccessId: process.env.AWS_BUCKET_ACCESS_ID,
     awsBucketToken: process.env.AWS_BUCKET_TOKEN,
     awsRegion: process.env.AWS_REGION,
     UploadFileUrlLink: process.env.UPLOAD_FILE_URL,
-    awsBucketName: process.env.AWS_BUCKET_NAME
+    awsBucketName: process.env.AWS_BUCKET_NAME,
+    SPOONACULAR_API_KEY: process.env.SPOONACULAR_API_KEY,
 }
