@@ -1,41 +1,18 @@
 // src/util/route_util.js
 
 import React from 'react';
-import { connect } from 'react-redux';
-import { Route, Redirect, withRouter } from 'react-router-dom';
+import {useSelector} from 'react-redux';
+import {Navigate} from 'react-router-dom';
 
 // Passed in from parent component or from mapStateToProps
-const Auth = ({ component: Component, path, loggedIn, exact }) => (
-    <Route path={path} exact={exact} render={(props) => (
-        !loggedIn ? (
-            <Component {...props} />
-        ) : (
-                // Redirect to the tweets page if the user is authenticated
-                <Redirect to="/" />
-            )
-    )} />
-);
+export const AuthRoute = ({ element }) => {
+  const loggedIn = useSelector(state => state.session.isAuthenticated);
+  return !loggedIn ? element : <Navigate to="/" />;
+};
 
-const Protected = ({ component: Component, loggedIn, ...rest }) => (
-    <Route
-        {...rest}
-        render={props =>
-            loggedIn ? (
-                <Component {...props} />
-            ) : (
-                    // Redirect to the login page if the user is already authenticated
-                    <Redirect to="/login" />
-                )
-        }
-    />
-);
+export const ProtectedRoute = ({ element }) => {
+  const loggedIn = useSelector(state => state.session.isAuthenticated);
+  return loggedIn ? element : <Navigate to="/login" />;
+};
 
 // Use the isAuthenitcated slice of state to determine whether a user is logged in
-
-const mapStateToProps = state => (
-    { loggedIn: state.session.isAuthenticated }
-);
-
-export const AuthRoute = withRouter(connect(mapStateToProps)(Auth));
-
-export const ProtectedRoute = withRouter(connect(mapStateToProps)(Protected));
