@@ -2,4 +2,4 @@ const validText = str => {
     return typeof str === 'string' && str.trim().length > 0;
 }
 
-module.exports = validText
+export default validText;
