@@ -1,19 +1,19 @@
 import React from 'react'
 import { Provider } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
-import { HashRouter, Route } from 'react-router-dom'
+import {HashRouter, Route, Routes} from 'react-router-dom'
 import App from './App'
 import Hamburger from "./modal/Hamburger";
 
 const Root = ({store, persistor}) => {
     return (
         <Provider store={store}>
-            <PersistGate persistor={persistor}>
+            {/*<PersistGate persistor={persistor}>*/}
                 <HashRouter>
-                    <Route path='/' component={App}/>
+                    <App/>
                     <Hamburger/>
                 </HashRouter>
-            </PersistGate>
+            {/*</PersistGate>*/}
         </Provider>
     )
 }
