@@ -1,7 +1,6 @@
 import { connect } from 'react-redux';
 import BottomNavBar from './BottomNavBar';
 import { openModal } from '../../actions/ModalActions';
-import {withRouter} from 'react-router-dom'
 
 const mapStateToProps = state => ({
     modal: state.ui.modal
@@ -11,4 +10,4 @@ const mapDispatchToProps = dispatch => ({
     openNewRoutineModal: () => dispatch(openModal('addRoutine')),
 })
 
-export default withRouter(connect(mapStateToProps, mapDispatchToProps)(BottomNavBar))
+export default connect(mapStateToProps, mapDispatchToProps)(BottomNavBar)
